requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.6.0
flask-requests==0.0.14
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python==4.5.1.48
pampy==0.3.0
Pillow==8.1.1
pkg-resources==0.0.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3
